Enable servicemp3 patch for all amlogic devices:
- HAVE_AMLOGIC is for all
- HAVE_ALIEN5 is for alien5 and mecool devices (for now)
- HAVE_MECOOL is for mecool devices only

If we found something that is not shared with alien5 for mecool then we need to use HAVE_MECOOL instead of HAVE_ALIEN5 everywhere, for now enigma2 is ok.

And guess what, PLi forgot to add HAVE_AMLOGIC to their servicemp3's configure.ac so till now we couldn't benefit from HAVE_AMLOGIC hackes.